creates profile embed wrapper for previewing sections
